+++ src/cobib/config/config.py
@@ -100,6 +100,9 @@ class Config(Dict[str, Any]):
             "bibtex": {
                 "ignore_non_standard_types": False,
             },
+            "yaml": {
+                "use_c_lib_yaml": False,
+            },
         },
         "tui": {
             "default_list_args": ["-l"],
@@ -510,6 +513,15 @@ class Config(Dict[str, Any]):
             isinstance(self.parsers.bibtex.ignore_non_standard_types, bool),
             "config.parsers.bibtex.ignore_non_standard_types should be a boolean.",
         )
+        self._assert(
+            isinstance(self.parsers.yaml.use_c_lib_yaml, bool),
+            "config.parsers.yaml.use_c_lib_yaml should be a boolean.",
+        )
+        if self.parsers.yaml.use_c_lib_yaml is False:
+            LOGGER.warning(
+                "The config.parsers.yaml.use_c_lib_yaml setting (introduced in version 3.4.0) will "
+                "change its default value to `True` in version 4.0.0!"
+            )
 
         # TUI section
         self._assert(